serialization and shared_ptr updates

+++ MW_API_Types.hpp
@@ -18,6 +18,6 @@ typedef enum MASC {
 	Failure
 } MW_API_STATUS_CODE;
 
-typedef unsigned long long MW_ID;
+typedef unsigned long MW_ID;
 
 #endif
